Refactor and pepping all over
