+++ requirements.txt
@@ -33,7 +33,10 @@ pyparsing==2.3.1
 python-dateutil==2.7.5
 pytz==2018.9
 requests==2.21.0
+scikit-learn==0.20.3
+scipy==1.2.1
 six==1.12.0
+sklearn==0.0
 snowballstemmer==1.2.1
 Sphinx==1.8.3
 sphinx-rtd-theme==0.4.2
